Implement moving room shares
When a recipient moves or deletes a room share the share should be moved
or deleted only for that recipient and be left untouched for the other
users in the room. Therefore, when a recipient moves or deletes a share
a copy of the original share is added to keep the share status
(permissions (to mark a share as deleted) and file target) from the
point of view of the recipient; when the share is later accessed by the
recipient RoomShareProvider returns it as seen by that recipient.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>